requirements.txt
azure-identity==1.12.0
msgraph-core==0.2.2
click==8.1.3
jmespath==1.0.1
tabulate==0.8.10
pygments==2.13.0
